+++ AceptaAdmin.php
@@ -10,52 +10,8 @@ if(isset($_SESSION['TxtUsr'])){
   
 }
 ?>
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
-  <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1.0"/>
-  <title>Contratación</title>
+<?php require 'php/header.php'; ?>
 
-  <!-- CSS  -->
-  <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-  <link href="css/materialize.css" type="text/css" rel="stylesheet" media="screen,projection"/>
-  <link href="css/style.css" type="text/css" rel="stylesheet" media="screen,projection"/>
-
-</head>
-<body class="grey darken-fondo2">
-  <?php
-    if(isset($_SESSION['TxtUsr'])){
-      require 'php/Dropdown.php';
-    }
-  ?>
-  <nav class="grey darken-4" role="navigation">
-    <div class="nav-wrapper container"><a id="logo-container" href="index.html" class="brand-logo"><img src="imgs/logoP4.png"  id="logoP"></a>
-      <ul class="right hide-on-med-and-down">
-        <?php
-          if(isset($_SESSION['TxtUsr'])){
-            require 'php/Bienvenido.php';
-          }
-        ?>
-        <li><a href="contratacion.html">Contratación</a></li>
-        <li><a href="comprobante.html">Comprobante</a></li>
-        <li><a href="admin.html">Admin</a></li>
-        
-      </ul>
-      
-      <ul id="nav-mobile" class="sidenav">
-      <?php
-          if(isset($_SESSION['TxtUsr'])){
-            echo "<li><a href='php/cerrarSesion.php'>CerrarSesion</a></li>";
-          }
-        ?>
-        <li><a href="contratacion.html">Contratación</a></li>
-        <li><a href="comprobante.html">Comprobante</a></li>
-        <li><a href="admin.html">Admin</a></li>
-      </ul>
-      <a href="#" data-target="nav-mobile" class="sidenav-trigger"><i class="material-icons">menu</i></a>
-    </div>
-  </nav>
   <div class="grey darken-vacio">
 
     <div class="container">
@@ -78,48 +34,4 @@ if(isset($_SESSION['TxtUsr'])){
   </div>
 
 
-  <footer class="page-footer grey darken-fondo2">
-    <div class="container">
-      <div class="row">
-        <div class="col l12">
-          <h5 class="white-text">Inserte nombre de la Empresa</h5>
-          <p class="grey-text text-lighten-4">Inserte un buen slogan aquí</p>
-
-
-        </div>
-        <div class="col l3 s12">
-          <h5 class="white-text"> Enlaces</h5>
-          <ul>
-            <li><a class="white-text" href="#!">Link 1</a></li>
-            <li><a class="white-text" href="#!">Link 2</a></li>
-            <li><a class="white-text" href="#!">Link 3</a></li>
-            <li><a class="white-text" href="#!">Link 4</a></li>
-          </ul>
-        </div>
-        <div class="col l3 s12">
-          <h5 class="white-text">Contactos</h5>
-          <ul>
-            <li><a class="white-text" href="#!">Link 1</a></li>
-            <li><a class="white-text" href="#!">Link 2</a></li>
-            <li><a class="white-text" href="#!">Link 3</a></li>
-            <li><a class="white-text" href="#!">Link 4</a></li>
-          </ul>
-        </div>
-      </div>
-    </div>
-    <div class="footer-copyright">
-      <div class="container">
-      Hecho en <a class="orange-text text-lighten-3" href="https://escom.ipn.mx">ESCOM</a>
-      </div>
-    </div>
-  </footer>
-
-
-  <!--  Scripts-->
-  <script src="https://code.jquery.com/jquery-2.1.1.min.js"></script>
-  <script src="js/materialize.js"></script>
-  <script src="js/init.js"></script>
-  <script src="js/admin.js"></script>
-
-  </body>
-</html>
+  <?php require 'php/footer.php'; ?>
