+++ index.php
@@ -1,35 +1,6 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
-  <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1.0"/>
-  <title>◾ DJ’s Hub LTD◾</title>
-
-  <!-- CSS  -->
-  <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-  <link href="css/materialize.css" type="text/css" rel="stylesheet" media="screen,projection"/>
-  <link href="css/style.css" type="text/css" rel="stylesheet" media="screen,projection"/>
-</head>
-<body class="grey darken-fondo2">
-  <nav class="grey darken-4" role="navigation">
-    <div class="nav-wrapper container"><a id="logo-container" href="index.html" class="brand-logo"><img class="grey darken-4 " src="imgs/logo_FINAL.png"  id="logoP"></a>
-      <ul class="right hide-on-med-and-down">
-        <li><a href="contratacion.html">Contratación</a></li>
-        <li><a href="comprobante.html">Comprobante</a></li>
-        <li><a href="admin.html">Admin</a></li>
-        
-      </ul>
-      
-      <ul id="nav-mobile" class="sidenav">
-        <li><a href="contratacion.html">Contratación</a></li>
-        <li><a href="comprobante.html">Comprobante</a></li>
-        <li><a href="admin.html">Admin</a></li>
-      </ul>
-      <a href="#" data-target="nav-mobile" class="sidenav-trigger"><i class="material-icons">menu</i></a>
-    </div>
-  </nav>
+<?php require 'php/header.php';?>
   
-  <div id="index-banner" class="grey darken-fondocorrido">
+<div id="index-banner" class="grey darken-fondocorrido">
     <div class="white-text darken-1 container">
       <br><br>
   
@@ -325,75 +296,4 @@
   </div>
 </div>
 
-<footer class="page-footer grey darken-fondo2">
-  <div class="container">
-    <div class="row">
-      <div class="col l6 s12">
-        <h5 class="white-text">Sobre nosotros</h5>
-        <p class="grey-text text-lighten-4">Somos un grupo de estudiantes de ESCOM realizando este proyecto para la materia "Desarrollo de Tecnologías para la Web".</p>
-        
-        
-      </div>
-      <div class="col l3 s12">
-        <h5 class="white-text">Integrantes</h5>
-        <ul>
-          <li><a class="white-text" href="#!">Arturo</a></li>
-          <li><a class="white-text" href="#!">Iván</a></li>
-            <li><a class="white-text" href="#!">Luis</a></li>
-            <li><a class="white-text" href="#!">Mario</a></li>
-          </ul>
-        </div>
-        <div class="col l3 s12">
-          <h5 class="white-text">GitHub</h5>
-          <ul>
-            <li>
-              <div class="col s2">
-                <img src="imgs/OldRevan66.png" class="circle responsive-img">
-              </div>
-              <div class="col s10">
-                <a class="white-text" href="https://github.com/OldRevan66">OldRevan66</a>
-              </div>
-            </li>
-            <li>
-              <div class="col s2">
-              <img src="imgs/ivanalejandro2002.png" class="circle responsive-img">
-              </div>
-              <div class="col s10">
-                <a class="white-text" href="https://github.com/ivanalejandro2002">ivanalejandro2002</a>
-              </div>
-            </li>
-            <li>
-              <div class="col s2">
-                <img src="imgs/Wuiszho.jpeg" class="circle responsive-img">
-              </div>
-              <div class="col s10">
-                <a class="white-text" href="https://github.com/Wuiszho">Wuiszho</a>
-              </div>
-            </li>
-            <li>
-              <div class="col s2">
-                <img src="imgs/MarioARC21.jpeg" class="circle responsive-img">
-              </div>
-              <div class="col s10">
-                <a class="white-text" href="https://github.com/MarioARC21">MarioARC21</a>
-              </div>
-            </li>
-          </ul>
-        </div>
-      </div>
-    </div>
-    <div class="footer-copyright">
-      <div class="container">
-      Made by <a class="orange-text text-lighten-3" href="http://materializecss.com">Materialize</a>
-      </div>
-    </div>
-  </footer>
-
-
-  <!--  Scripts-->
-  <script src="https://code.jquery.com/jquery-2.1.1.min.js"></script>
-  <script src="js/materialize.js"></script>
-  <script src="js/init.js"></script>
-
-  </body>
-</html>
+<?php require 'php/footer.php';?>
