+++ php/footer.php
@@ -1,0 +1,75 @@
+    <footer class="page-footer grey darken-fondo2">
+      <div class="container">
+        <div class="row">
+          <div class="col l6 s12">
+            <h5 class="white-text">Sobre nosotros</h5>
+            <p class="grey-text text-lighten-4">Somos un grupo de estudiantes de ESCOM realizando este proyecto para la materia "Desarrollo de Tecnologías para la Web".</p>
+            
+            
+          </div>
+          <div class="col l3 s12">
+            <h5 class="white-text">Integrantes</h5>
+            <ul>
+              <li><a class="white-text" href="#!">Arturo</a></li>
+              <li><a class="white-text" href="#!">Iván</a></li>
+              <li><a class="white-text" href="#!">Luis</a></li>
+              <li><a class="white-text" href="#!">Mario</a></li>
+            </ul>
+            </div>
+              <div class="col l3 s12">
+                <h5 class="white-text">GitHub</h5>
+                <ul>
+                  <li>
+                    <div class="col s2">
+                      <img src="imgs/OldRevan66.png" class="circle responsive-img">
+                    </div>
+                    <div class="col s10">
+                      <a class="white-text" href="https://github.com/OldRevan66">OldRevan66</a>
+                    </div>
+                  </li>
+                  <li>
+                    <div class="col s2">
+                      <img src="imgs/ivanalejandro2002.png" class="circle responsive-img">
+                      </div>
+                      <div class="col s10">
+                        <a class="white-text" href="https://github.com/ivanalejandro2002">ivanalejandro2002</a>
+                    </div>
+                  </li>
+                  <li>
+                    <div class="col s2">
+                      <img src="imgs/Wuiszho.jpeg" class="circle responsive-img">
+                    </div>
+                    <div class="col s10">
+                      <a class="white-text" href="https://github.com/Wuiszho">Wuiszho</a>
+                    </div>
+                  </li>
+                  <li>
+                    <div class="col s2">
+                      <img src="imgs/MarioARC21.jpeg" class="circle responsive-img">
+                    </div>
+                    <div class="col s10">
+                      <a class="white-text" href="https://github.com/MarioARC21">MarioARC21</a>
+                    </div>
+                  </li>
+                </ul>
+              </div>
+            </div>
+          </div>
+        <div class="footer-copyright">
+          <div class="container">
+            Made by <a class="orange-text text-lighten-3" href="http://materializecss.com">Materialize</a>
+          </div>
+        </div>
+      </div>
+    </footer>
+
+
+    <!--  Scripts-->
+    <script src="https://code.jquery.com/jquery-2.1.1.min.js"></script>
+    <script src="js/materialize.js"></script>
+    <script src="js/init.js"></script>
+    <script src="js/contratacion.js"></script>
+    <script src="js/admin.js"></script>
+
+  </body>
+</html>
