+++ php/header.php
@@ -1,0 +1,3 @@
+<?php require 'php/inicioH.php';
+
+ require 'php/finH.php';?>
